chore(dsp-transfer-process-dispatcher): refactor rebase
